feat :: useNewProduct 작성
